engines/requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
accelerate>=0.20.0
Pillow>=10.0.0 
diffusers>=0.19.0
opencv-python>=4.7.0
protobuf>=4.25.1 